Boundary checks in simulation window
Added a way to check for the boundaries of the simulation window, so that particles bounce of the walls....
Perhaps there shouldn't be bounds and I should be able to drag and scroll sideways and up/down to see the stuff outside... no idea, but something to think about.

I'm thinking I should change the way in which particles are created, I'm thinking about click to select the point at which the particle is created and then drag to a given point to give the particle a speed... some sort of spring-like mechanism, when the mouse is released, the particle's speed is set.