Began work to replace the system in simulation.py
Began working on a more stable and general particle physics system
based on the paper "Physically Based Modeling" by Andrew Witkin of
Pixar.

I opted for this route after spending some time with the traer.physics
library for Processing. Traer has put together a nice system which
handles the physics very well and it's not concerned with displaying
the particles per se (that should be something the user of the system
would take care of). While his system is now more than just what is
described in Witkin's paper, I think I would be better off starting
with the paper and then evolving it as the need arises.

I have not yet finished it though as there are several things to do:

	-implement the euler_step function
	-implement forces, springs, and the like
	-create a simple vector class for the force, position, and velocity
	 vectors in the particle class (I could use Euclid, but I would like
	 this to be somewhat of a standalone library of sorts... we'll see)
	-Test it thoroughly.

I will continue this tomorrow during my lunch hour or something.

--R.